move legacy migrations to separate file
