+++ packages/tlschema/src/records/TLAsset.ts
@@ -1,13 +1,9 @@
-import { createRecordType, defineMigrations, RecordId } from '@tldraw/store'
+import { createRecordType, RecordId } from '@tldraw/store'
 import { T } from '@tldraw/validate'
 import { TLBaseAsset } from '../assets/TLBaseAsset'
-import {
-	bookmarkAssetMigrations,
-	bookmarkAssetValidator,
-	TLBookmarkAsset,
-} from '../assets/TLBookmarkAsset'
-import { imageAssetMigrations, imageAssetValidator, TLImageAsset } from '../assets/TLImageAsset'
-import { TLVideoAsset, videoAssetMigrations, videoAssetValidator } from '../assets/TLVideoAsset'
+import { bookmarkAssetValidator, TLBookmarkAsset } from '../assets/TLBookmarkAsset'
+import { imageAssetValidator, TLImageAsset } from '../assets/TLImageAsset'
+import { TLVideoAsset, videoAssetValidator } from '../assets/TLVideoAsset'
 import { TLShape } from './TLShape'
 
 /** @public */
@@ -22,38 +18,6 @@ export const assetValidator: T.Validator<TLAsset> = T.model(
 		bookmark: bookmarkAssetValidator,
 	})
 )
-
-/** @internal */
-export const assetVersions = {
-	AddMeta: 1,
-}
-
-/** @internal */
-// eslint-disable-next-line deprecation/deprecation
-export const assetMigrations = defineMigrations({
-	subTypeKey: 'type',
-	subTypeMigrations: {
-		image: imageAssetMigrations,
-		video: videoAssetMigrations,
-		bookmark: bookmarkAssetMigrations,
-	},
-	currentVersion: assetVersions.AddMeta,
-	migrators: {
-		[assetVersions.AddMeta]: {
-			up: (record) => {
-				return {
-					...record,
-					meta: {},
-				}
-			},
-			down: ({ meta: _, ...record }) => {
-				return {
-					...record,
-				}
-			},
-		},
-	},
-})
 
 /** @public */
 export type TLAssetPartial<T extends TLAsset = TLAsset> = T extends T
