+++ packages/tlschema/src/records/TLCamera.ts
@@ -1,4 +1,4 @@
-import { BaseRecord, createRecordType, defineMigrations, RecordId } from '@tldraw/store'
+import { BaseRecord, createRecordType, RecordId } from '@tldraw/store'
 import { JsonObject } from '@tldraw/utils'
 import { T } from '@tldraw/validate'
 import { idValidator } from '../misc/id-validator'
@@ -33,32 +33,6 @@ export const cameraValidator: T.Validator<TLCamera> = T.model(
 		meta: T.jsonValue as T.ObjectValidator<JsonObject>,
 	})
 )
-
-/** @internal */
-export const cameraVersions = {
-	AddMeta: 1,
-}
-
-/** @internal */
-// eslint-disable-next-line deprecation/deprecation
-export const cameraMigrations = defineMigrations({
-	currentVersion: cameraVersions.AddMeta,
-	migrators: {
-		[cameraVersions.AddMeta]: {
-			up: (record) => {
-				return {
-					...record,
-					meta: {},
-				}
-			},
-			down: ({ meta: _, ...record }) => {
-				return {
-					...record,
-				}
-			},
-		},
-	},
-})
 
 /** @public */
 export const CameraRecordType = createRecordType<TLCamera>('camera', {
