+++ packages/tlschema/src/records/TLInstance.ts
@@ -1,4 +1,4 @@
-import { BaseRecord, createRecordType, defineMigrations, RecordId } from '@tldraw/store'
+import { BaseRecord, createRecordType, RecordId } from '@tldraw/store'
 import { JsonObject } from '@tldraw/utils'
 import { T } from '@tldraw/validate'
 import { BoxModel, boxModelValidator } from '../misc/geometry-types'
@@ -159,387 +159,6 @@ export function createInstanceRecordType(stylesById: Map<string, StyleProp<unkno
 		})
 	)
 }
-
-/** @internal */
-export const instanceVersions = {
-	AddTransparentExportBgs: 1,
-	RemoveDialog: 2,
-	AddToolLockMode: 3,
-	RemoveExtraPropsForNextShape: 4,
-	AddLabelColor: 5,
-	AddFollowingUserId: 6,
-	RemoveAlignJustify: 7,
-	AddZoom: 8,
-	AddVerticalAlign: 9,
-	AddScribbleDelay: 10,
-	RemoveUserId: 11,
-	AddIsPenModeAndIsGridMode: 12,
-	HoistOpacity: 13,
-	AddChat: 14,
-	AddHighlightedUserIds: 15,
-	ReplacePropsForNextShapeWithStylesForNextShape: 16,
-	AddMeta: 17,
-	RemoveCursorColor: 18,
-	AddLonelyProperties: 19,
-	ReadOnlyReadonly: 20,
-	AddHoveringCanvas: 21,
-	AddScribbles: 22,
-	AddInset: 23,
-	AddDuplicateProps: 24,
-} as const
-
-/** @public */
-// eslint-disable-next-line deprecation/deprecation
-export const instanceMigrations = defineMigrations({
-	currentVersion: instanceVersions.AddDuplicateProps,
-	migrators: {
-		[instanceVersions.AddTransparentExportBgs]: {
-			up: (instance: TLInstance) => {
-				return { ...instance, exportBackground: true }
-			},
-			down: ({ exportBackground: _, ...instance }: TLInstance) => {
-				return instance
-			},
-		},
-		[instanceVersions.RemoveDialog]: {
-			up: ({ dialog: _, ...instance }: any) => {
-				return instance
-			},
-			down: (instance: TLInstance) => {
-				return { ...instance, dialog: null }
-			},
-		},
-		[instanceVersions.AddToolLockMode]: {
-			up: (instance: TLInstance) => {
-				return { ...instance, isToolLocked: false }
-			},
-			down: ({ isToolLocked: _, ...instance }: TLInstance) => {
-				return instance
-			},
-		},
-		[instanceVersions.RemoveExtraPropsForNextShape]: {
-			up: ({ propsForNextShape, ...instance }: any) => {
-				return {
-					...instance,
-					propsForNextShape: Object.fromEntries(
-						Object.entries(propsForNextShape).filter(([key]) =>
-							[
-								'color',
-								'labelColor',
-								'dash',
-								'fill',
-								'size',
-								'font',
-								'align',
-								'verticalAlign',
-								'icon',
-								'geo',
-								'arrowheadStart',
-								'arrowheadEnd',
-								'spline',
-							].includes(key)
-						)
-					),
-				}
-			},
-			down: (instance: TLInstance) => {
-				// we can't restore these, so do nothing :/
-				return instance
-			},
-		},
-		[instanceVersions.AddLabelColor]: {
-			up: ({ propsForNextShape, ...instance }: any) => {
-				return {
-					...instance,
-					propsForNextShape: {
-						...propsForNextShape,
-						labelColor: 'black',
-					},
-				}
-			},
-			down: (instance) => {
-				const { labelColor: _, ...rest } = instance.propsForNextShape
-				return {
-					...instance,
-					propsForNextShape: {
-						...rest,
-					},
-				}
-			},
-		},
-		[instanceVersions.AddFollowingUserId]: {
-			up: (instance: TLInstance) => {
-				return { ...instance, followingUserId: null }
-			},
-			down: ({ followingUserId: _, ...instance }: TLInstance) => {
-				return instance
-			},
-		},
-		[instanceVersions.RemoveAlignJustify]: {
-			up: (instance: any) => {
-				let newAlign = instance.propsForNextShape.align
-				if (newAlign === 'justify') {
-					newAlign = 'start'
-				}
-
-				return {
-					...instance,
-					propsForNextShape: {
-						...instance.propsForNextShape,
-						align: newAlign,
-					},
-				}
-			},
-			down: (instance: TLInstance) => {
-				return { ...instance }
-			},
-		},
-		[instanceVersions.AddZoom]: {
-			up: (instance: TLInstance) => {
-				return { ...instance, zoomBrush: null }
-			},
-			down: ({ zoomBrush: _, ...instance }: TLInstance) => {
-				return instance
-			},
-		},
-		[instanceVersions.AddVerticalAlign]: {
-			up: (instance) => {
-				return {
-					...instance,
-					propsForNextShape: {
-						...instance.propsForNextShape,
-						verticalAlign: 'middle',
-					},
-				}
-			},
-			down: (instance) => {
-				const { verticalAlign: _, ...propsForNextShape } = instance.propsForNextShape
-				return {
-					...instance,
-					propsForNextShape,
-				}
-			},
-		},
-		[instanceVersions.AddScribbleDelay]: {
-			up: (instance) => {
-				if (instance.scribble !== null) {
-					return { ...instance, scribble: { ...instance.scribble, delay: 0 } }
-				}
-				return { ...instance }
-			},
-			down: (instance) => {
-				if (instance.scribble !== null) {
-					const { delay: _delay, ...rest } = instance.scribble
-					return { ...instance, scribble: rest }
-				}
-				return { ...instance }
-			},
-		},
-		[instanceVersions.RemoveUserId]: {
-			up: ({ userId: _, ...instance }: any) => {
-				return instance
-			},
-			down: (instance: TLInstance) => {
-				return { ...instance, userId: 'user:none' }
-			},
-		},
-		[instanceVersions.AddIsPenModeAndIsGridMode]: {
-			up: (instance: TLInstance) => {
-				return { ...instance, isPenMode: false, isGridMode: false }
-			},
-			down: ({ isPenMode: _, isGridMode: __, ...instance }: TLInstance) => {
-				return instance
-			},
-		},
-		[instanceVersions.HoistOpacity]: {
-			up: ({ propsForNextShape: { opacity, ...propsForNextShape }, ...instance }: any) => {
-				return { ...instance, opacityForNextShape: Number(opacity ?? '1'), propsForNextShape }
-			},
-			down: ({ opacityForNextShape: opacity, ...instance }: any) => {
-				return {
-					...instance,
-					propsForNextShape: {
-						...instance.propsForNextShape,
-						opacity:
-							opacity < 0.175
-								? '0.1'
-								: opacity < 0.375
-									? '0.25'
-									: opacity < 0.625
-										? '0.5'
-										: opacity < 0.875
-											? '0.75'
-											: '1',
-					},
-				}
-			},
-		},
-		[instanceVersions.AddChat]: {
-			up: (instance: TLInstance) => {
-				return { ...instance, chatMessage: '', isChatting: false }
-			},
-			down: ({ chatMessage: _, isChatting: __, ...instance }: TLInstance) => {
-				return instance
-			},
-		},
-		[instanceVersions.AddHighlightedUserIds]: {
-			up: (instance: TLInstance) => {
-				return { ...instance, highlightedUserIds: [] }
-			},
-			down: ({ highlightedUserIds: _, ...instance }: TLInstance) => {
-				return instance
-			},
-		},
-		[instanceVersions.ReplacePropsForNextShapeWithStylesForNextShape]: {
-			up: ({ propsForNextShape: _, ...instance }) => {
-				return { ...instance, stylesForNextShape: {} }
-			},
-			down: ({ stylesForNextShape: _, ...instance }: TLInstance) => {
-				return {
-					...instance,
-					propsForNextShape: {
-						color: 'black',
-						labelColor: 'black',
-						dash: 'draw',
-						fill: 'none',
-						size: 'm',
-						icon: 'file',
-						font: 'draw',
-						align: 'middle',
-						verticalAlign: 'middle',
-						geo: 'rectangle',
-						arrowheadStart: 'none',
-						arrowheadEnd: 'arrow',
-						spline: 'line',
-					},
-				}
-			},
-		},
-		[instanceVersions.AddMeta]: {
-			up: (record) => {
-				return {
-					...record,
-					meta: {},
-				}
-			},
-			down: ({ meta: _, ...record }) => {
-				return {
-					...record,
-				}
-			},
-		},
-		[instanceVersions.RemoveCursorColor]: {
-			up: (record) => {
-				const { color: _, ...cursor } = record.cursor
-				return {
-					...record,
-					cursor,
-				}
-			},
-			down: (record) => {
-				return {
-					...record,
-					cursor: {
-						...record.cursor,
-						color: 'black',
-					},
-				}
-			},
-		},
-		[instanceVersions.AddLonelyProperties]: {
-			up: (record) => {
-				return {
-					...record,
-					canMoveCamera: true,
-					isFocused: false,
-					devicePixelRatio: 1,
-					isCoarsePointer: false,
-					openMenus: [],
-					isChangingStyle: false,
-					isReadOnly: false,
-				}
-			},
-			down: ({
-				canMoveCamera: _canMoveCamera,
-				isFocused: _isFocused,
-				devicePixelRatio: _devicePixelRatio,
-				isCoarsePointer: _isCoarsePointer,
-				openMenus: _openMenus,
-				isChangingStyle: _isChangingStyle,
-				isReadOnly: _isReadOnly,
-				...record
-			}) => {
-				return {
-					...record,
-				}
-			},
-		},
-		[instanceVersions.ReadOnlyReadonly]: {
-			up: ({ isReadOnly: _isReadOnly, ...record }) => {
-				return {
-					...record,
-					isReadonly: _isReadOnly,
-				}
-			},
-			down: ({ isReadonly: _isReadonly, ...record }) => {
-				return {
-					...record,
-					isReadOnly: _isReadonly,
-				}
-			},
-		},
-		[instanceVersions.AddHoveringCanvas]: {
-			up: (record) => {
-				return {
-					...record,
-					isHoveringCanvas: null,
-				}
-			},
-			down: ({ isHoveringCanvas: _, ...record }) => {
-				return {
-					...record,
-				}
-			},
-		},
-		[instanceVersions.AddScribbles]: {
-			up: ({ scribble: _, ...record }) => {
-				return {
-					...record,
-					scribbles: [],
-				}
-			},
-			down: ({ scribbles: _, ...record }) => {
-				return { ...record, scribble: null }
-			},
-		},
-		[instanceVersions.AddInset]: {
-			up: (record) => {
-				return {
-					...record,
-					insets: [false, false, false, false],
-				}
-			},
-			down: ({ insets: _, ...record }) => {
-				return {
-					...record,
-				}
-			},
-		},
-		[instanceVersions.AddDuplicateProps]: {
-			up: (record) => {
-				return {
-					...record,
-					duplicateProps: null,
-				}
-			},
-			down: ({ duplicateProps: _, ...record }) => {
-				return {
-					...record,
-				}
-			},
-		},
-	},
-})
 
 /** @public */
 export const TLINSTANCE_ID = 'instance:instance' as TLInstanceId
