+++ packages/tlschema/src/records/TLPageState.ts
@@ -1,10 +1,8 @@
-import { BaseRecord, createRecordType, defineMigrations, RecordId } from '@tldraw/store'
+import { BaseRecord, createRecordType, RecordId } from '@tldraw/store'
 import { JsonObject } from '@tldraw/utils'
 import { T } from '@tldraw/validate'
 import { idValidator } from '../misc/id-validator'
 import { shapeIdValidator } from '../shapes/TLBaseShape'
-import { CameraRecordType } from './TLCamera'
-import { TLINSTANCE_ID } from './TLInstance'
 import { pageIdValidator, TLPage } from './TLPage'
 import { TLShapeId } from './TLShape'
 
@@ -45,152 +43,6 @@ export const instancePageStateValidator: T.Validator<TLInstancePageState> = T.mo
 		meta: T.jsonValue as T.ObjectValidator<JsonObject>,
 	})
 )
-
-/** @internal */
-export const instancePageStateVersions = {
-	AddCroppingId: 1,
-	RemoveInstanceIdAndCameraId: 2,
-	AddMeta: 3,
-	RenameProperties: 4,
-	RenamePropertiesAgain: 5,
-} as const
-
-/** @public */
-// eslint-disable-next-line deprecation/deprecation
-export const instancePageStateMigrations = defineMigrations({
-	currentVersion: instancePageStateVersions.RenamePropertiesAgain,
-	migrators: {
-		[instancePageStateVersions.AddCroppingId]: {
-			up(instance) {
-				return { ...instance, croppingShapeId: null }
-			},
-			down({ croppingShapeId: _croppingShapeId, ...instance }) {
-				return instance
-			},
-		},
-		[instancePageStateVersions.RemoveInstanceIdAndCameraId]: {
-			up({ instanceId: _, cameraId: __, ...instance }) {
-				return instance
-			},
-			down(instance) {
-				// this should never be called since we bump the schema version
-				return {
-					...instance,
-					instanceId: TLINSTANCE_ID,
-					cameraId: CameraRecordType.createId('void'),
-				}
-			},
-		},
-		[instancePageStateVersions.AddMeta]: {
-			up: (record) => {
-				return {
-					...record,
-					meta: {},
-				}
-			},
-			down: ({ meta: _, ...record }) => {
-				return {
-					...record,
-				}
-			},
-		},
-		[instancePageStateVersions.RenameProperties]: {
-			// this migration is cursed: it was written wrong and doesn't do anything.
-			// rather than replace it, I've added another migration below that fixes it.
-			up: (record) => {
-				const {
-					selectedShapeIds,
-					hintingShapeIds,
-					erasingShapeIds,
-					hoveredShapeId,
-					editingShapeId,
-					croppingShapeId,
-					focusedGroupId,
-					...rest
-				} = record
-				return {
-					selectedShapeIds: selectedShapeIds,
-					hintingShapeIds: hintingShapeIds,
-					erasingShapeIds: erasingShapeIds,
-					hoveredShapeId: hoveredShapeId,
-					editingShapeId: editingShapeId,
-					croppingShapeId: croppingShapeId,
-					focusedGroupId: focusedGroupId,
-					...rest,
-				}
-			},
-			down: (record) => {
-				const {
-					selectedShapeIds,
-					hintingShapeIds,
-					erasingShapeIds,
-					hoveredShapeId,
-					editingShapeId,
-					croppingShapeId,
-					focusedGroupId,
-					...rest
-				} = record
-				return {
-					selectedShapeIds: selectedShapeIds,
-					hintingShapeIds: hintingShapeIds,
-					erasingShapeIds: erasingShapeIds,
-					hoveredShapeId: hoveredShapeId,
-					editingShapeId: editingShapeId,
-					croppingShapeId: croppingShapeId,
-					focusedGroupId: focusedGroupId,
-					...rest,
-				}
-			},
-		},
-		[instancePageStateVersions.RenamePropertiesAgain]: {
-			up: (record) => {
-				const {
-					selectedIds,
-					hintingIds,
-					erasingIds,
-					hoveredId,
-					editingId,
-					croppingShapeId,
-					croppingId,
-					focusLayerId,
-					...rest
-				} = record
-				return {
-					...rest,
-					selectedShapeIds: selectedIds,
-					hintingShapeIds: hintingIds,
-					erasingShapeIds: erasingIds,
-					hoveredShapeId: hoveredId,
-					editingShapeId: editingId,
-					croppingShapeId: croppingShapeId ?? croppingId ?? null,
-					focusedGroupId: focusLayerId,
-				}
-			},
-			down: (record) => {
-				const {
-					selectedShapeIds,
-					hintingShapeIds,
-					erasingShapeIds,
-					hoveredShapeId,
-					editingShapeId,
-					croppingShapeId,
-					focusedGroupId,
-					...rest
-				} = record
-				return {
-					...rest,
-					selectedIds: selectedShapeIds,
-					hintingIds: hintingShapeIds,
-					erasingIds: erasingShapeIds,
-					hoveredId: hoveredShapeId,
-					editingId: editingShapeId,
-					croppingId: croppingShapeId,
-					focusLayerId: focusedGroupId,
-				}
-			},
-		},
-	},
-})
 
 /** @public */
 export const InstancePageStateRecordType = createRecordType<TLInstancePageState>(
