+++ packages/tlschema/src/records/TLPointer.ts
@@ -1,4 +1,4 @@
-import { BaseRecord, createRecordType, defineMigrations, RecordId } from '@tldraw/store'
+import { BaseRecord, createRecordType, RecordId } from '@tldraw/store'
 import { JsonObject } from '@tldraw/utils'
 import { T } from '@tldraw/validate'
 import { idValidator } from '../misc/id-validator'
@@ -30,32 +30,6 @@ export const pointerValidator: T.Validator<TLPointer> = T.model(
 		meta: T.jsonValue as T.ObjectValidator<JsonObject>,
 	})
 )
-
-/** @internal */
-export const pointerVersions = {
-	AddMeta: 1,
-}
-
-/** @internal */
-// eslint-disable-next-line deprecation/deprecation
-export const pointerMigrations = defineMigrations({
-	currentVersion: pointerVersions.AddMeta,
-	migrators: {
-		[pointerVersions.AddMeta]: {
-			up: (record) => {
-				return {
-					...record,
-					meta: {},
-				}
-			},
-			down: ({ meta: _, ...record }) => {
-				return {
-					...record,
-				}
-			},
-		},
-	},
-})
 
 /** @public */
 export const PointerRecordType = createRecordType<TLPointer>('pointer', {
