+++ packages/tlschema/src/records/TLPresence.ts
@@ -1,11 +1,10 @@
-import { BaseRecord, createRecordType, defineMigrations, RecordId } from '@tldraw/store'
+import { BaseRecord, createRecordType, RecordId } from '@tldraw/store'
 import { JsonObject } from '@tldraw/utils'
 import { T } from '@tldraw/validate'
 import { BoxModel, boxModelValidator } from '../misc/geometry-types'
 import { idValidator } from '../misc/id-validator'
 import { cursorTypeValidator, TLCursor } from '../misc/TLCursor'
 import { scribbleValidator, TLScribble } from '../misc/TLScribble'
-import { TLINSTANCE_ID } from './TLInstance'
 import { TLPageId } from './TLPage'
 import { TLShapeId } from './TLShape'
 
@@ -66,82 +65,6 @@ export const instancePresenceValidator: T.Validator<TLInstancePresence> = T.mode
 		meta: T.jsonValue as T.ObjectValidator<JsonObject>,
 	})
 )
-
-/** @internal */
-export const instancePresenceVersions = {
-	AddScribbleDelay: 1,
-	RemoveInstanceId: 2,
-	AddChatMessage: 3,
-	AddMeta: 4,
-	RenameSelectedShapeIds: 5,
-} as const
-
-// eslint-disable-next-line deprecation/deprecation
-export const instancePresenceMigrations = defineMigrations({
-	currentVersion: instancePresenceVersions.RenameSelectedShapeIds,
-	migrators: {
-		[instancePresenceVersions.AddScribbleDelay]: {
-			up: (instance) => {
-				if (instance.scribble !== null) {
-					return { ...instance, scribble: { ...instance.scribble, delay: 0 } }
-				}
-				return { ...instance }
-			},
-			down: (instance) => {
-				if (instance.scribble !== null) {
-					const { delay: _delay, ...rest } = instance.scribble
-					return { ...instance, scribble: rest }
-				}
-				return { ...instance }
-			},
-		},
-		[instancePresenceVersions.RemoveInstanceId]: {
-			up: ({ instanceId: _, ...instance }) => {
-				return instance
-			},
-			down: (instance) => {
-				return { ...instance, instanceId: TLINSTANCE_ID }
-			},
-		},
-		[instancePresenceVersions.AddChatMessage]: {
-			up: (instance) => {
-				return { ...instance, chatMessage: '' }
-			},
-			down: ({ chatMessage: _, ...instance }) => {
-				return instance
-			},
-		},
-		[instancePresenceVersions.AddMeta]: {
-			up: (record) => {
-				return {
-					...record,
-					meta: {},
-				}
-			},
-			down: ({ meta: _, ...record }) => {
-				return {
-					...record,
-				}
-			},
-		},
-		[instancePresenceVersions.RenameSelectedShapeIds]: {
-			up: (record) => {
-				const { selectedShapeIds, ...rest } = record
-				return {
-					selectedShapeIds: selectedShapeIds,
-					...rest,
-				}
-			},
-			down: (record) => {
-				const { selectedShapeIds, ...rest } = record
-				return {
-					selectedShapeIds: selectedShapeIds,
-					...rest,
-				}
-			},
-		},
-	},
-})
 
 /** @public */
 export const InstancePresenceRecordType = createRecordType<TLInstancePresence>(
